refactor(CostRequestView): 구입금액을 요청하는 뷰를 클래스로 분리
View 로직이 너무 무겁고 복잡해지는 문제가 있어서, 입출력을 담당하는 비즈니스 로직을 기준으로 클래스를 분리한다.
